Move generalMobilePlatform under src/common
